requirements.txt
pendulum == 3.1.0
pandas
matplotlib
pytest==8.3.3
pre-commit==4.0.1
ruff==0.6.9
sqlalchemy
psycopg2-binary
apache-airflow-providers-postgres==5.2.0
pymongo==4.13.2
